backend/requirements.txt
argon2-cffi==23.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
django-health-check==3.20.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
